[ZF3][Behat] added CV feature tests
