+++ behat.yml.dist
@@ -49,3 +49,5 @@ default:
                 - Yawik\Behat\Select2Context
                 - Yawik\Behat\JobContext
                 - Yawik\Behat\ApplicationContext
+                - Yawik\Behat\SummaryFormContext
+                - Yawik\Behat\CvContext
